Adjust "build-static-bins.sh" to allow for more selective building (so we don't have to rebuild containerd and runc every time we modify Docker patches)

+++ build-static-bins.sh
@@ -4,15 +4,29 @@ set -e
 cd "$(dirname "$(readlink -f "$BASH_SOURCE")")"
 
 targetDir="$PWD/static-bins"
-rm -rf "$targetDir"
-mkdir "$targetDir"
 
-parts=( docker containerd runc )
-snapcraft clean --step pull "${parts[@]}"
-snapcraft pull "${parts[@]}"
+# usage: $0 part [part ...]
+#    ie: $0 docker
+if [ "$#" -lt 1 ]; then
+	set -- docker containerd runc
+	rm -rf "$targetDir" # if we're building all three, let's do it clean
+fi
+parts=( "$@" )
+
+mkdir -p "$targetDir"
+
+# make an easy lookup table
+declare -A wantPart=()
+for part; do
+	wantPart[$part]=1
+done
+
+( set -x && snapcraft clean --step pull "${parts[@]}" )
+( set -x && snapcraft pull "${parts[@]}" )
 
 export SNAPDIR="$PWD"
 
+if [ "${wantPart[docker]}" ]; then
 (
 	cd parts/docker/src
 
@@ -20,15 +34,17 @@ export SNAPDIR="$PWD"
 
 	# devicemapper and static builds don't get along well
 	export DOCKER_BUILDTAGS="$DOCKER_BUILDTAGS exclude_graphdriver_devicemapper"
-	./hack/make.sh binary
+	( set -x && ./hack/make.sh binary )
 
 	dockerBin='bundles/latest/binary/docker'
-	"$dockerBin" -v
+	( set -x && "$dockerBin" -v )
 	"$dockerBin" -v | grep -q "$DOCKER_GITCOMMIT"
 
 	install -T "$(readlink -f "$dockerBin")" "$targetDir/docker"
 )
+fi
 
+if [ "${wantPart[containerd]}" ]; then
 (
 	cd parts/containerd/src
 
@@ -36,15 +52,17 @@ export SNAPDIR="$PWD"
 	ln -sfT "$PWD" .gopath/src/github.com/docker/containerd
 	export GOPATH="$PWD/.gopath"
 
-	make static GIT_COMMIT= GIT_BRANCH= LDFLAGS=
+	( set -x && make static GIT_COMMIT= GIT_BRANCH= LDFLAGS= )
 
-	bin/containerd -v
+	( set -x && bin/containerd -v )
 
 	install -T bin/containerd "$targetDir/docker-containerd"
 	install -T bin/containerd-shim "$targetDir/docker-containerd-shim"
 	install -T bin/ctr "$targetDir/docker-containerd-ctr"
 )
+fi
 
+if [ "${wantPart[runc]}" ]; then
 (
 	cd parts/runc/src
 
@@ -52,9 +70,10 @@ export SNAPDIR="$PWD"
 	ln -sfT "$PWD" .gopath/src/github.com/opencontainers/runc
 	export GOPATH="$PWD/.gopath"
 
-	make static BUILDTAGS='seccomp apparmor selinux' COMMIT=
+	( set -x && make static BUILDTAGS='seccomp apparmor selinux' COMMIT= )
 
-	./runc -v
+	( set -x && ./runc -v )
 
 	install -T runc "$targetDir/docker-runc"
 )
+fi
